+++ kernel/include/Luna/misc/font.hpp
@@ -3,7 +3,7 @@
 #include <Luna/common.hpp>
 
 namespace font {
-    extern uint8_t width;
-    extern uint8_t height;
+    extern int32_t width;
+    extern int32_t height;
     extern uint8_t bitmap[];
 } // namespace font
